+++ cmd/monetd/commands/run.go
@@ -60,6 +60,7 @@ func bindFlags(cmd *cobra.Command) {
 	cmd.Flags().Bool("babble.bootstrap", configuration.Global.Babble.Bootstrap, "bootstrap Babble from database")
 	cmd.Flags().String("babble.moniker", configuration.Global.Babble.Moniker, "friendly name")
 	cmd.Flags().Bool("babble.maintenance-mode", configuration.Global.Babble.MaintenanceMode, "start babble in suspended (non-gossipping) state")
+	cmd.Flags().Int("babble.suspend-limit", configuration.Global.Babble.SuspendLimit, "number of undetermined-events since last run that will trigger automatic suspension")
 
 	// Eth config
 	cmd.Flags().Int("eth.cache", configuration.Global.Eth.Cache, "megabytes of memory allocated to internal caching (min 16MB / database forced)")